deployments/prompt_llava/requirements.txt
--extra-index-url=https://download.pytorch.org/whl/
accelerate==0.30.1
bitsandbytes==0.43.1
keras==2.15.0
Pillow==9.4.0
requests==2.31.0
tensorflow==2.15.0
torch==2.3.0+cu121
transformers==4.41.2